+++ cli/tauri.js/src/types/config.validator.ts
@@ -287,6 +287,22 @@ export const TauriConfigSchema = {
       ],
       "description": "tauri root configuration object",
       "properties": {
+        "allowlist": {
+          "additionalProperties": {
+            "type": "boolean"
+          },
+          "defaultProperties": [
+          ],
+          "properties": {
+            "all": {
+              "type": "boolean"
+            }
+          },
+          "required": [
+            "all"
+          ],
+          "type": "object"
+        },
         "bundle": {
           "additionalProperties": false,
           "defaultProperties": [
@@ -401,17 +417,6 @@ export const TauriConfigSchema = {
           "$ref": "#/definitions/CliConfig",
           "description": "app's CLI definition"
         },
-        "edge": {
-          "additionalProperties": false,
-          "defaultProperties": [
-          ],
-          "properties": {
-            "active": {
-              "type": "boolean"
-            }
-          },
-          "type": "object"
-        },
         "embeddedServer": {
           "additionalProperties": false,
           "defaultProperties": [
@@ -461,22 +466,6 @@ export const TauriConfigSchema = {
           },
           "type": "object"
         },
-        "whitelist": {
-          "additionalProperties": {
-            "type": "boolean"
-          },
-          "defaultProperties": [
-          ],
-          "properties": {
-            "all": {
-              "type": "boolean"
-            }
-          },
-          "required": [
-            "all"
-          ],
-          "type": "object"
-        },
         "window": {
           "additionalProperties": false,
           "defaultProperties": [
@@ -505,12 +494,11 @@ export const TauriConfigSchema = {
         }
       },
       "required": [
+        "allowlist",
         "bundle",
-        "edge",
         "embeddedServer",
         "inliner",
         "security",
-        "whitelist",
         "window"
       ],
       "type": "object"