+++ tauri/src/endpoints.rs
@@ -1,16 +1,12 @@
 mod cmd;
 #[allow(unused_imports)]
 mod file_system;
-mod init;
 mod salt;
-
-use init::init;
 
 #[cfg(assets)]
 mod asset;
 #[cfg(open)]
 mod browser;
-#[cfg(any(open_dialog, save_dialog))]
 mod dialog;
 #[cfg(event)]
 mod event;
@@ -19,24 +15,15 @@ mod http;
 #[cfg(notification)]
 mod notification;
 
-use web_view::WebView;
+use webview_official::Webview;
 
 #[allow(unused_variables)]
-pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> crate::Result<()> {
+pub(crate) fn handle(webview: &mut Webview, arg: &str) -> crate::Result<()> {
   use cmd::Cmd::*;
   match serde_json::from_str(arg) {
     Err(e) => Err(e.into()),
     Ok(command) => {
       match command {
-        Init {} => {
-          let event_init = init()?;
-          webview.eval(&format!(
-            r#"{event_init}
-                window.external.invoke('{{"cmd":"__initialized"}}')
-              "#,
-            event_init = event_init
-          ))?;
-        }
         ReadTextFile {
           path,
           options,
@@ -46,7 +33,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(read_text_file)]
           file_system::read_text_file(webview, path, options, callback, error);
           #[cfg(not(read_text_file))]
-          whitelist_error(webview, error, "readTextFile");
+          allowlist_error(webview, error, "readTextFile");
         }
         ReadBinaryFile {
           path,
@@ -57,7 +44,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(read_binary_file)]
           file_system::read_binary_file(webview, path, options, callback, error);
           #[cfg(not(read_binary_file))]
-          whitelist_error(webview, error, "readBinaryFile");
+          allowlist_error(webview, error, "readBinaryFile");
         }
         WriteFile {
           path,
@@ -69,7 +56,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(write_file)]
           file_system::write_file(webview, path, contents, options, callback, error);
           #[cfg(not(write_file))]
-          whitelist_error(webview, error, "writeFile");
+          allowlist_error(webview, error, "writeFile");
         }
         WriteBinaryFile {
           path,
@@ -81,7 +68,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(write_binary_file)]
           file_system::write_binary_file(webview, path, contents, options, callback, error);
           #[cfg(not(write_binary_file))]
-          whitelist_error(webview, error, "writeBinaryFile");
+          allowlist_error(webview, error, "writeBinaryFile");
         }
         ReadDir {
           path,
@@ -92,7 +79,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(read_dir)]
           file_system::read_dir(webview, path, options, callback, error);
           #[cfg(not(read_dir))]
-          whitelist_error(webview, error, "readDir");
+          allowlist_error(webview, error, "readDir");
         }
         CopyFile {
           source,
@@ -104,7 +91,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(copy_file)]
           file_system::copy_file(webview, source, destination, options, callback, error);
           #[cfg(not(copy_file))]
-          whitelist_error(webview, error, "copyFile");
+          allowlist_error(webview, error, "copyFile");
         }
         CreateDir {
           path,
@@ -115,7 +102,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(create_dir)]
           file_system::create_dir(webview, path, options, callback, error);
           #[cfg(not(create_dir))]
-          whitelist_error(webview, error, "createDir");
+          allowlist_error(webview, error, "createDir");
         }
         RemoveDir {
           path,
@@ -126,7 +113,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(remove_dir)]
           file_system::remove_dir(webview, path, options, callback, error);
           #[cfg(not(remove_dir))]
-          whitelist_error(webview, error, "removeDir");
+          allowlist_error(webview, error, "removeDir");
         }
         RemoveFile {
           path,
@@ -137,7 +124,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(remove_file)]
           file_system::remove_file(webview, path, options, callback, error);
           #[cfg(not(remove_file))]
-          whitelist_error(webview, error, "removeFile");
+          allowlist_error(webview, error, "removeFile");
         }
         RenameFile {
           old_path,
@@ -149,13 +136,13 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(rename_file)]
           file_system::rename_file(webview, old_path, new_path, options, callback, error);
           #[cfg(not(rename_file))]
-          whitelist_error(webview, error, "renameFile");
+          allowlist_error(webview, error, "renameFile");
         }
         SetTitle { title } => {
           #[cfg(set_title)]
-          webview.set_title(&title)?;
+          webview.set_title(&title);
           #[cfg(not(set_title))]
-          throw_whitelist_error(webview, "title");
+          throw_allowlist_error(webview, "title");
         }
         Execute {
           command,
@@ -166,13 +153,13 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(execute)]
           crate::call(webview, command, args, callback, error);
           #[cfg(not(execute))]
-          throw_whitelist_error(webview, "execute");
+          throw_allowlist_error(webview, "execute");
         }
         Open { uri } => {
           #[cfg(open)]
           browser::open(uri);
           #[cfg(not(open))]
-          throw_whitelist_error(webview, "open");
+          throw_allowlist_error(webview, "open");
         }
         ValidateSalt {
           salt,
@@ -189,16 +176,16 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(event)]
           {
             let js_string = event::listen_fn(event, handler, once)?;
-            webview.eval(&js_string)?;
+            webview.eval(&js_string);
           }
           #[cfg(not(event))]
-          throw_whitelist_error(webview, "event");
+          throw_allowlist_error(webview, "event");
         }
         Emit { event, payload } => {
           #[cfg(event)]
           crate::event::on_event(event, payload);
           #[cfg(not(event))]
-          throw_whitelist_error(webview, "event");
+          throw_allowlist_error(webview, "event");
         }
         OpenDialog {
           options,
@@ -208,7 +195,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(open_dialog)]
           dialog::open(webview, options, callback, error)?;
           #[cfg(not(open_dialog))]
-          whitelist_error(webview, error, "title");
+          allowlist_error(webview, error, "title");
         }
         SaveDialog {
           options,
@@ -218,7 +205,38 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(save_dialog)]
           dialog::save(webview, options, callback, error)?;
           #[cfg(not(save_dialog))]
-          throw_whitelist_error(webview, "saveDialog");
+          throw_allowlist_error(webview, "saveDialog");
+        }
+        MessageDialog { message } => {
+          let exe = std::env::current_exe()?;
+          let exe_dir = exe.parent().expect("failed to get exe directory");
+          let app_name = exe
+            .file_name()
+            .expect("failed to get exe filename")
+            .to_string_lossy();
+          dialog::message(app_name.to_string(), message);
+        }
+        AskDialog {
+          title,
+          message,
+          callback,
+          error,
+        } => {
+          let exe = std::env::current_exe()?;
+          dialog::ask(
+            webview,
+            title.unwrap_or_else(|| {
+              let exe_dir = exe.parent().expect("failed to get exe directory");
+              exe
+                .file_name()
+                .expect("failed to get exe filename")
+                .to_string_lossy()
+                .to_string()
+            }),
+            message,
+            callback,
+            error,
+          )?;
         }
         HttpRequest {
           options,
@@ -228,7 +246,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(http_request)]
           http::make_request(webview, *options, callback, error);
           #[cfg(not(http_request))]
-          whitelist_error(webview, error, "httpRequest");
+          allowlist_error(webview, error, "httpRequest");
         }
         #[cfg(assets)]
         LoadAsset {
@@ -254,7 +272,7 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           api_error(
             webview,
             error,
-            "CLI definition not set under tauri.conf.json > tauri > cli (https://tauri.studio/docs/api/config#tauri)",
+            "CLI definition not set under tauri.conf.json > tauri > cli (https://tauri.studio/docs/api/config#tauri.cli)",
           );
         }
         Notification {
@@ -265,19 +283,19 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
           #[cfg(notification)]
           notification::send(webview, options, callback, error);
           #[cfg(not(notification))]
-          whitelist_error(webview, error, "notification");
+          allowlist_error(webview, error, "notification");
         }
         IsNotificationPermissionGranted { callback, error } => {
           #[cfg(notification)]
           notification::is_permission_granted(webview, callback, error);
           #[cfg(not(notification))]
-          whitelist_error(webview, error, "notification");
+          allowlist_error(webview, error, "notification");
         }
         RequestNotificationPermission { callback, error } => {
           #[cfg(notification)]
           notification::request_permission(webview, callback, error)?;
           #[cfg(not(notification))]
-          whitelist_error(webview, error, "notification");
+          allowlist_error(webview, error, "notification");
         }
       }
       Ok(())
@@ -286,58 +304,35 @@ pub(crate) fn handle<T: 'static>(webview: &mut WebView<'_, T>, arg: &str) -> cra
 }
 
 #[allow(dead_code)]
-fn api_error<T: 'static>(webview: &mut WebView<'_, T>, error_fn: String, message: &str) {
+fn api_error(webview: &mut Webview, error_fn: String, message: &str) {
   let reject_code = tauri_api::rpc::format_callback(error_fn, message);
-  webview
-    .eval(&reject_code)
-    .expect("failed to eval api error")
+  webview.eval(&reject_code)
 }
 
 #[allow(dead_code)]
-fn whitelist_error<T: 'static>(
-  webview: &mut WebView<'_, T>,
-  error_fn: String,
-  whitelist_key: &str,
-) {
+fn allowlist_error(webview: &mut Webview, error_fn: String, allowlist_key: &str) {
   api_error(
     webview,
     error_fn,
     &format!(
-      "{}' not whitelisted (https://tauri.studio/docs/api/config#tauri)",
-      whitelist_key
+      "{}' not on the allowlist (https://tauri.studio/docs/api/config#tauri.allowlist)",
+      allowlist_key
     ),
   )
 }
 
 #[allow(dead_code)]
-fn throw_whitelist_error<T: 'static>(webview: &mut WebView<'_, T>, whitelist_key: &str) {
-  let reject_code = format!(r#"throw new Error("'{}' not whitelisted")"#, whitelist_key);
-  webview
-    .eval(&reject_code)
-    .expect("failed to eval whitelist error")
+fn throw_allowlist_error(webview: &mut Webview, allowlist_key: &str) {
+  let reject_code = format!(
+    r#"throw new Error("'{}' not on the allowlist")"#,
+    allowlist_key
+  );
+  webview.eval(&reject_code)
 }
 
 #[cfg(test)]
 mod test {
   use proptest::prelude::*;
-
-  #[test]
-  // test to see if check init produces a string or not.
-  fn check_init() {
-    if cfg!(not(event)) {
-      let res = super::init();
-      match res {
-        Ok(s) => assert_eq!(s, ""),
-        Err(e) => panic!("init Err {:?}", e.to_string()),
-      }
-    } else if cfg!(event) {
-      let res = super::init();
-      match res {
-        Ok(s) => assert!(s.contains("window.__TAURI__.promisified")),
-        Err(e) => panic!("init Err {:?}", e.to_string()),
-      }
-    }
-  }
 
   // check the listen_fn for various usecases.
   proptest! {