+++ tauri/src/endpoints/dialog.rs
@@ -1,9 +1,13 @@
 use super::cmd::{OpenDialogOptions, SaveDialogOptions};
-use crate::api::dialog::{pick_folder, save_file, select, select_multiple, Response};
+use crate::api::dialog::{
+  ask as ask_dialog, message as message_dialog, pick_folder, save_file, select, select_multiple,
+  DialogSelection, Response,
+};
 use serde_json::Value as JsonValue;
-use web_view::WebView;
+use webview_official::Webview;
 
 /// maps a dialog response to a JS value to eval
+#[cfg(any(open_dialog, save_dialog))]
 fn map_response(response: Response) -> JsonValue {
   match response {
     Response::Okay(path) => path.into(),
@@ -14,8 +18,8 @@ fn map_response(response: Response) -> JsonValue {
 
 /// Shows an open dialog.
 #[cfg(open_dialog)]
-pub fn open<T: 'static>(
-  webview: &mut WebView<'_, T>,
+pub fn open(
+  webview: &mut Webview,
   options: OpenDialogOptions,
   callback: String,
   error: String,
@@ -40,8 +44,8 @@ pub fn open<T: 'static>(
 
 /// Shows a save dialog.
 #[cfg(save_dialog)]
-pub fn save<T: 'static>(
-  webview: &mut WebView<'_, T>,
+pub fn save(
+  webview: &mut Webview,
   options: SaveDialogOptions,
   callback: String,
   error: String,
@@ -49,6 +53,31 @@ pub fn save<T: 'static>(
   crate::execute_promise_sync(
     webview,
     move || save_file(options.filter, options.default_path).map(map_response),
+    callback,
+    error,
+  )?;
+  Ok(())
+}
+
+/// Shows a message in a dialog.
+pub fn message(title: String, message: String) {
+  message_dialog(message, title);
+}
+
+/// Shows a dialog with a yes/no question.
+pub fn ask(
+  webview: &mut Webview,
+  title: String,
+  message: String,
+  callback: String,
+  error: String,
+) -> crate::Result<()> {
+  crate::execute_promise_sync(
+    webview,
+    move || match ask_dialog(message, title) {
+      DialogSelection::Yes => Ok(true),
+      _ => Ok(false),
+    },
     callback,
     error,
   )?;