+++ tauri/src/event.rs
@@ -6,7 +6,7 @@ use lazy_static::lazy_static;
 use once_cell::sync::Lazy;
 use serde::Serialize;
 use serde_json::Value as JsonValue;
-use web_view::Handle;
+use webview_official::WebviewMut;
 
 /// An event handler.
 struct EventHandler {
@@ -57,8 +57,8 @@ pub fn listen<F: FnMut(Option<String>) + Send + 'static>(id: impl Into<String>, 
 }
 
 /// Emits an event to JS.
-pub fn emit<T: 'static, S: Serialize>(
-  webview_handle: &Handle<T>,
+pub fn emit<S: Serialize>(
+  webview: &mut WebviewMut,
   event: impl AsRef<str> + Send + 'static,
   payload: Option<S>,
 ) -> crate::Result<()> {
@@ -70,17 +70,15 @@ pub fn emit<T: 'static, S: Serialize>(
     JsonValue::Null
   };
 
-  webview_handle
-    .dispatch(move |_webview| {
-      _webview.eval(&format!(
-        "window['{}']({{type: '{}', payload: {}}}, '{}')",
-        emit_function_name(),
-        event.as_ref(),
-        js_payload,
-        salt
-      ))
-    })
-    .expect("Failed to dispatch JS from emit");
+  webview.dispatch(move |webview_ref| {
+    webview_ref.eval(&format!(
+      "window['{}']({{type: '{}', payload: {}}}, '{}')",
+      emit_function_name(),
+      event.as_ref(),
+      js_payload,
+      salt
+    ))
+  })?;
 
   Ok(())
 }